# Improve shuffling performance