--- a/ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/helpers/MiscHelpers.java
+++ b/ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/helpers/MiscHelpers.java
@@ -20,12 +20,10 @@ import static tech.pegasys.teku.spec.logic.common.helpers.MathHelpers.uintToByte
 
 import com.google.common.primitives.UnsignedBytes;
 import it.unimi.dsi.fastutil.ints.IntList;
-import java.security.MessageDigest;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
-import tech.pegasys.teku.infrastructure.crypto.Hash;
-import tech.pegasys.teku.infrastructure.crypto.MessageDigestFactory;
+import tech.pegasys.teku.infrastructure.crypto.Sha256;
 import tech.pegasys.teku.infrastructure.ssz.Merkleizable;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteVector;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
@@ -41,9 +39,9 @@ public class MiscHelpers {
   // Math.pow(2, 8) - 1;
   public static final UInt64 MAX_RANDOM_BYTE = UInt64.valueOf(255);
 
-  static final byte[] EMPTY_HASH = Bytes.EMPTY.toArrayUnsafe();
-
   protected final SpecConfig specConfig;
+
+  private static final byte[] EMPTY_HASH = Bytes.EMPTY.toArrayUnsafe();
 
   public MiscHelpers(final SpecConfig specConfig) {
     this.specConfig = specConfig;
@@ -51,6 +49,7 @@ public class MiscHelpers {
 
   public int computeShuffledIndex(int index, int indexCount, Bytes32 seed) {
     checkArgument(index < indexCount, "CommitteeUtil.computeShuffledIndex1");
+    final Sha256 sha256 = new Sha256();
 
     int indexRet = index;
     final int shuffleRoundCount = specConfig.getShuffleRoundCount();
@@ -61,17 +60,17 @@ public class MiscHelpers {
 
       // This needs to be unsigned modulo.
       int pivot =
-          bytesToUInt64(Hash.sha256(Bytes.wrap(seed, roundAsByte)).slice(0, 8))
+          bytesToUInt64(sha256.wrappedDigest(seed, roundAsByte).slice(0, 8))
               .mod(indexCount)
               .intValue();
       int flip = Math.floorMod(pivot + indexCount - indexRet, indexCount);
       int position = Math.max(indexRet, flip);
 
       Bytes positionDiv256 = uintToBytes(Math.floorDiv(position, 256L), 4);
-      Bytes hashBytes = Hash.sha256(seed, roundAsByte, positionDiv256);
+      byte[] hashBytes = sha256.digest(seed, roundAsByte, positionDiv256);
 
       int bitIndex = position & 0xff;
-      int theByte = hashBytes.get(bitIndex / 8);
+      int theByte = hashBytes[bitIndex / 8];
       int theBit = (theByte >> (bitIndex & 0x07)) & 1;
       if (theBit != 0) {
         indexRet = flip;
@@ -83,15 +82,16 @@ public class MiscHelpers {
 
   public int computeProposerIndex(BeaconState state, IntList indices, Bytes32 seed) {
     checkArgument(!indices.isEmpty(), "compute_proposer_index indices must not be empty");
+    final Sha256 sha256 = new Sha256();
     int i = 0;
     final int total = indices.size();
-    Bytes32 hash = null;
+    byte[] hash = null;
     while (true) {
       int candidateIndex = indices.getInt(computeShuffledIndex(i % total, total, seed));
       if (i % 32 == 0) {
-        hash = Hash.sha256(seed, uint64ToBytes(Math.floorDiv(i, 32L)));
+        hash = sha256.digest(seed, uint64ToBytes(Math.floorDiv(i, 32L)));
       }
-      int randomByte = UnsignedBytes.toInt(hash.get(i % 32));
+      int randomByte = UnsignedBytes.toInt(hash[i % 32]);
       UInt64 effectiveBalance = state.getValidators().get(candidateIndex).getEffectiveBalance();
       if (effectiveBalance
           .times(MAX_RANDOM_BYTE)
@@ -164,7 +164,7 @@ public class MiscHelpers {
   }
 
   public void shuffleList(int[] input, Bytes32 seed) {
-    final MessageDigest digest = MessageDigestFactory.createSha256();
+    final Sha256 sha256 = new Sha256();
 
     int listSize = input.length;
     if (listSize == 0) {
@@ -173,12 +173,11 @@ public class MiscHelpers {
 
     for (int round = specConfig.getShuffleRoundCount() - 1; round >= 0; round--) {
 
-      Bytes roundAsByte = Bytes.of((byte) round);
+      final Bytes roundAsByte = Bytes.of((byte) round);
 
       // This needs to be unsigned modulo.
-      seed.update(digest);
-      roundAsByte.update(digest);
-      int pivot = bytesToUInt64(Bytes.wrap(digest.digest()).slice(0, 8)).mod(listSize).intValue();
+      final Bytes hash = sha256.wrappedDigest(seed, roundAsByte);
+      int pivot = bytesToUInt64(hash.slice(0, 8)).mod(listSize).intValue();
 
       byte[] hashBytes = EMPTY_HASH;
       int mirror1 = (pivot + 2) / 2;
@@ -190,21 +189,13 @@ public class MiscHelpers {
           flip = pivot - i;
           bitIndex = i & 0xff;
           if (bitIndex == 0 || i == mirror1) {
-            seed.update(digest);
-            roundAsByte.update(digest);
-            uintToBytes(i / 256, 4).update(digest);
-
-            hashBytes = digest.digest();
+            hashBytes = sha256.digest(seed, roundAsByte, uintToBytes(i / 256, 4));
           }
         } else {
           flip = pivot + listSize - i;
           bitIndex = flip & 0xff;
           if (bitIndex == 0xff || i == pivot + 1) {
-            seed.update(digest);
-            roundAsByte.update(digest);
-            uintToBytes(flip / 256, 4).update(digest);
-
-            hashBytes = digest.digest();
+            hashBytes = sha256.digest(seed, roundAsByte, uintToBytes(flip / 256, 4));
           }
         }