Avoids multiple `createSha256` and `Bytes.wrap` by introducing a `Sha256` class designed to be reused.


Before:
```
# Run complete. Total time: 00:01:26

Benchmark                            (indexCount)  Mode  Cnt  Score   Error  Units
ShuffleBenchmark.shuffledIndexBench         16384    ss   15  1.088 ± 0.017   s/op
ShuffleBenchmark.shuffledIndexBench         32768    ss   15  2.283 ± 0.185   s/op
ShuffleBenchmark.shuffledListBench          16384    ss   15  0.012 ± 0.004   s/op
ShuffleBenchmark.shuffledListBench          32768    ss   15  0.026 ± 0.003   s/op
```

After:
```
# Run complete. Total time: 00:01:17

Benchmark                            (indexCount)  Mode  Cnt  Score    Error  Units
ShuffleBenchmark.shuffledIndexBench         16384    ss   15  0.992 ±  0.020   s/op
ShuffleBenchmark.shuffledIndexBench         32768    ss   15  1.958 ±  0.021   s/op
ShuffleBenchmark.shuffledListBench          16384    ss   15  0.008 ±  0.001   s/op
ShuffleBenchmark.shuffledListBench          32768    ss   15  0.014 ±  0.001   s/op
```

Memory allocations comparison on shuffle benchmark:

<img width="1371" alt="image" src="https://user-images.githubusercontent.com/15999009/189854702-381b9485-54be-4c37-b916-54c2ee4a3eaf.png">

<img width="1373" alt="image" src="https://user-images.githubusercontent.com/15999009/189854818-50f26444-0b2b-4b4e-9394-5999ce76c496.png">


related to: #6016

## Documentation

- [x] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [x] I thought about adding a changelog entry, and added one if I deemed necessary.