@ajsutton 2022-09-12
Worth a changelog entry as well. Otherwise I agree with Paul about extracting a method to do the hash. Looks good otherwise.