@tbenr 2022-09-13
ouch, `computeProposerIndex` doesn't have a unit test