@tbenr 2022-09-13
I see I can apply the same pattern to:

`getNextSyncCommitteeIndices` and `isValidMerkleBranch`.

And maybe do some cleanup to ssz to use `Sha256` class